+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 4.0 GUI and Plugin Refactoring
 
 - 2023-??-?? 4.0.30
+    - GUI
+        - added battery to StatusBar
     - refactoring
         - moved StatusBarFeature to ctl
 - 2023-12-13 4.0.29